Создал модули: Data, Display, Simulation
