+++ .vs/CMakeWorkspaceSettings.json
@@ -1,0 +1,3 @@
+{
+  "enableCMake": false
+}
