+++ Display.h
@@ -1,0 +1,15 @@
+#pragma once
+#include <windows.h>
+#include <iostream>
+#include <vector>
+#include <GL/gl.h>
+#include <GL/glu.h>
+#include "GL/freeglut.h"
+#include <glm/glm.hpp>
+#include <glm/gtc/matrix_transform.hpp>
+#include <glm/gtc/type_ptr.hpp> 
+
+using namespace std;
+using namespace glm;
+
+void display(void);
