+++ Simulation.cpp
@@ -1,0 +1,7 @@
+#include "Simulation.h"
+#include "Data.h"
+
+void simulation(int value) {
+	glutPostRedisplay();
+	glutTimerFunc(mSIMULATION_TIME, simulation, 0);
+};
