+++ Simulation.h
@@ -1,0 +1,9 @@
+#pragma once
+#include <windows.h>
+#include <iostream>
+#include <vector>
+#include <GL/gl.h>
+#include <GL/glu.h>
+#include "GL/freeglut.h"
+
+void simulation(int value);
