+++ x64/Debug/Graphic.exe.recipe
@@ -2,7 +2,7 @@
 <Project>
   <ProjectOutputs>
     <ProjectOutput>
-      <FullPath>C:\Projects\AOKG\4\Graphic\x64\Debug\Graphic.exe</FullPath>
+      <FullPath>C:\STUDY\АОКГ\4\GL\Graphic\x64\Debug\Graphic.exe</FullPath>
     </ProjectOutput>
   </ProjectOutputs>
   <ContentFiles />
