+++ x64/Debug/Graphic.tlog/Cl.items.tlog
@@ -1,0 +1,6 @@
+C:\STUDY\АОКГ\4\GL\Graphic\Camera.cpp;C:\STUDY\АОКГ\4\GL\Graphic\x64\Debug\Camera.obj
+C:\STUDY\АОКГ\4\GL\Graphic\Data.cpp;C:\STUDY\АОКГ\4\GL\Graphic\x64\Debug\Data.obj
+C:\STUDY\АОКГ\4\GL\Graphic\Display.cpp;C:\STUDY\АОКГ\4\GL\Graphic\x64\Debug\Display.obj
+C:\STUDY\АОКГ\4\GL\Graphic\Graphic.cpp;C:\STUDY\АОКГ\4\GL\Graphic\x64\Debug\Graphic.obj
+C:\STUDY\АОКГ\4\GL\Graphic\Object.cpp;C:\STUDY\АОКГ\4\GL\Graphic\x64\Debug\Object.obj
+C:\STUDY\АОКГ\4\GL\Graphic\Simulation.cpp;C:\STUDY\АОКГ\4\GL\Graphic\x64\Debug\Simulation.obj
