+++ x64/Debug/Graphic.tlog/Graphic.lastbuildstate
@@ -1,2 +1,2 @@
-PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.31.31103:TargetPlatformVersion=10.0.19041.0:
-Debug|x64|C:\Projects\AOKG\4\Graphic\|
+PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.36.32532:TargetPlatformVersion=10.0.22000.0:
+Debug|x64|C:\STUDY\АОКГ\4\GL\Graphic\|
