+++ x64/Debug/Graphic.vcxproj.FileListAbsolute.txt
@@ -1,1 +1,2 @@
 C:\Projects\AOKG\4\Graphic\x64\Debug\Graphic.exe
+C:\STUDY\АОКГ\4\GL\Graphic\x64\Debug\Graphic.exe
